backend/requirements.txt
flask==3.1.2
ollama==0.6.0
python-dotenv==1.1.1